requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
matplotlib
numpy=1.26.4
pandas=2.2.2
python=3.9
pytorch=2.4.0=py3.9_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8
pytorch-mutex=1.0=cuda
scikit-learn=1.5.1
scipy
seaborn=0.13.2
setuptools=72.1.0
sip=6.7.12
six=1.16.0
tokenizers
torch-geometric
torchaudio=2.4.0=py39_cu118
torchtriton=3.0.0=py39
torchvision=0.19.0=py39_cu118
tornado=6.4.1
tqdm=4.66.5
transformers
